feat: Ahora se usa ApiResponse, mappers y exception handlers. Arreglados errores varios

+++ src/main/java/com/mza_agrotours/backend/dtos/reservas/ConsultarReservaDetalleDTO.java
@@ -1,11 +1,13 @@
 package com.mza_agrotours.backend.dtos.reservas;
 
+import java.math.BigDecimal;
+
 public record ConsultarReservaDetalleDTO(
         // ReservaDetalle - renglón, nombre, tipo rango etario, subtotal
         Integer renglon,
         String nombre,
         String tipoRangoEtario,
-        Float subtotal
+        BigDecimal subtotal
 ) {
 
 }
